Move GamemodeTabItem to a distinct class
